프로그래머스 2022 KAKAO BLIND RECRUITMENT k진수에서 소수 - 2차 풀이 실패

+++ CodingTest/Programmers/Solution.java
@@ -23,11 +23,11 @@ class Solution {
     private boolean isPrimeNum(final String s) {
         if (s.isBlank()) return false;
 
-        final int val = Integer.parseInt(s);
+        final long val = Long.parseLong(s);
 
         if (val <= 1) return false;
 
-        final int root = (int) Math.sqrt(val);
+        final long root = (long) Math.sqrt(val);
         for (int i = 2; i <= root; i++) {
             if (val % i == 0) {
                 return false;
